refactor: move searchitem to schemas
